Cast width argument for %-.*s formatting to int.

+++ transact.c
@@ -1088,7 +1088,7 @@ process_headers:
       ctl->server.envelope && !strcasecmp(ctl->server.envelope, "Delivered-To"))
    {
 	   if (outlevel >= O_DEBUG)
-		   report(stdout, GT_("Parsing envelope \"%s\" names \"%-.*s\"\n"), ctl->server.envelope, strcspn(delivered_to+2+strlen(ctl->server.envelope), "\r\n"), delivered_to+2+strlen(ctl->server.envelope));
+		   report(stdout, GT_("Parsing envelope \"%s\" names \"%-.*s\"\n"), ctl->server.envelope, (int)strcspn(delivered_to+2+strlen(ctl->server.envelope), "\r\n"), delivered_to+2+strlen(ctl->server.envelope));
 	    find_server_names(delivered_to, ctl, &msgblk.recipients);
 	    xfree(delivered_to);
    }
@@ -1101,7 +1101,7 @@ process_headers:
 	     * hostnames go through.
 	     */
 	   if (outlevel >= O_DEBUG)
-		   report(stdout, GT_("Parsing Received names \"%-.*s\"\n"), strcspn(received_for+2, "\r\n"), received_for+2);
+		   report(stdout, GT_("Parsing Received names \"%-.*s\"\n"), (int)strcspn(received_for+2, "\r\n"), received_for+2);
 	    find_server_names(received_for, ctl, &msgblk.recipients);
 	} else {
 	    /*
@@ -1128,7 +1128,7 @@ process_headers:
 	    /* now look for remaining adresses */
 	    while (to_addrchain) {
 		    if (outlevel >= O_DEBUG)
-			    report(stdout, GT_("Guessing from header \"%-.*s\".\n"), strcspn(msgblk.headers+to_addrchain->offset, "\r\n"), msgblk.headers+to_addrchain->offset);
+			    report(stdout, GT_("Guessing from header \"%-.*s\".\n"), (int)strcspn(msgblk.headers+to_addrchain->offset, "\r\n"), msgblk.headers+to_addrchain->offset);
 
 		find_server_names(msgblk.headers+to_addrchain->offset, ctl, &msgblk.recipients);
 		nextptr = to_addrchain->next;
